Remove Okapi
It's unmaintained, and doesn't look like it's ever coming back to life.

The final straw has been the annoyance of adding all the `allow-newer`s to projects which depend on anything in this one but don't even use Okapi. Though the fact that those are attempted to be built at all seems like a Cabal bug.